feat: add account group assignment functionality
- Implemented a new endpoint to update account group associations in accounts-routes.js.
- Added setAccountGroupSchema for validation in validators.js.
- Created setAccountGroup function in ledger-service.js to handle group assignments.
- Enhanced AccountsScreen.jsx to include UI for creating accounts and assigning them to groups.
- Updated GroupScreen.jsx to allow assigning existing accounts to groups and creating new accounts directly within the group context.
- Added reconnect logic in worker's index.js to handle database connection errors.
- Improved ModuleSidebar.jsx to use unique identifiers for navigation items.

+++ apps/api/src/routes/ledger/group-service.js
@@ -76,8 +76,8 @@ export function createGroupService({ prisma }) {
     const rows = await prisma.$queryRaw`
       SELECT g.*,
         gm.role AS my_role,
-        COUNT(DISTINCT gm2.user_id) FILTER (WHERE gm2.status = 'active') AS member_count,
-        COUNT(DISTINCT a.id) FILTER (WHERE a.enabled = true) AS account_count
+        COUNT(DISTINCT gm2.user_id) FILTER (WHERE gm2.status = 'active')::int4 AS member_count,
+        COUNT(DISTINCT a.id) FILTER (WHERE a.enabled = true)::int4 AS account_count
       FROM ledger_group g
       LEFT JOIN ledger_group_member gm
         ON gm.group_id = g.id AND gm.user_id = ${actorId}::uuid AND gm.status = 'active'
